easy_ocr_bn_en/requirements.txt
opencv-python
numpy
Pillow
torch==2.7.1
torchvision==0.22.1
git+https://github.com/csebuetnlp/normalizer
easyocr